style: cleanup and rename
Change name from packActivityDirectivesBothInPlan to just packActivityDirectivesInPlan. Removed unnecessary code and used functionality from code that already exists in Aerie.
